pyproject.toml
[project]
name = "ashareinsight"
version = "0.1.0"
description = "Enterprise concept retrieval system with PostgreSQL and vector search"
authors = [
    {name = "AShareInsight Team"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core Framework
    "langchain>=0.3.26",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-postgres>=0.0.12",
    "langgraph>=0.2.0",
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.25.0",
    # Database
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pgvector>=0.2.4",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # Caching
    "redis>=5.0.0",
    # Model Services
    "httpx>=0.25.0",
    "langchain-google-genai>=2.0.0",
    # Evaluation & Monitoring
    "ragas>=0.2.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.40b0",
    "prometheus-client>=0.19.0",
    # Utilities
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
    "akshare>=1.17.25",
    "chinese-calendar>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "ipython>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "YTT", "S", "B", "A", "C4", "T20"]
ignore = []

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.13"
# Start with less strict settings and gradually increase
warn_return_any = false
warn_unused_configs = true
no_implicit_reexport = true
namespace_packages = true
explicit_package_bases = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true
warn_incomplete_stub = false
# Focus on import errors and basic type safety
show_error_codes = true
pretty = true

# Per-module options
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "alembic.*"
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "langchain_community.*",
    "langchain_postgres.*",
    "langchain_google_genai.*",
    "sqlalchemy.*",
    "structlog.*",
    "tqdm.*",
    "click.*",
    "rich.*",
    "pytest.*",
    "opentelemetry.*",
    "prometheus_client.*",
    "pgvector.*",
    "ragas.*",
    "httpx.*",
    "redis.*",
    "uvicorn.*",
    "psycopg.*"
]
ignore_missing_imports = true
